pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.14.0",
    "asyncpg>=0.31.0",
    "datetime>=6.0",
    "dependency-injector==4.48.3",
    "fastapi>=0.124.4",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "mangum>=0.19.0",
    "psycopg>=3.3.2",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
